+++ core/src/schema/property/types/index.ts
@@ -4,3 +4,4 @@ export * from './SchemaDate';
 export * from './SchemaNumber';
 export * from './SchemaObject';
 export * from './SchemaString';
+export * from './SchemaFile';